requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.0.1
faiss-cpu==1.7.4
numpy==1.26.0
google-generativeai==0.3.1
pydantic>=2.11.8
pydantic-settings>=2.10.1
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
typing-extensions>=4.12.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
